+++ support-frontend/assets/helpers/subscriptionsForms/formValidation.ts
@@ -11,7 +11,6 @@ import {
 } from 'helpers/redux/checkout/address/validation';
 import type { SubscriptionsState } from 'helpers/redux/subscriptionsStore';
 import type { Action } from 'helpers/subscriptionsForms/formActions';
-// eslint-disable-next-line import/no-cycle -- these are quite tricky to unpick so we should come back to this
 import { setFormErrors } from 'helpers/subscriptionsForms/formActions';
 import type {
 	FormField,
