+++ support-frontend/assets/helpers/subscriptionsForms/submit.ts
@@ -42,13 +42,11 @@ import type { DirectDebitState } from 'helpers/redux/checkout/payment/directDebi
 import { getSubscriptionType } from 'helpers/redux/checkout/product/selectors/productType';
 import type { SubscriptionsState } from 'helpers/redux/subscriptionsStore';
 import type { Action } from 'helpers/subscriptionsForms/formActions';
-// eslint-disable-next-line import/no-cycle -- these are quite tricky to unpick so we should come back to this
 import {
 	setFormSubmitted,
 	setStage,
 	setSubmissionError,
 } from 'helpers/subscriptionsForms/formActions';
-// eslint-disable-next-line import/no-cycle -- these are quite tricky to unpick so we should come back to this
 import {
 	validateCheckoutForm,
 	validateWithDeliveryForm,
